fix(frontend): render one AI providers page on oss, ee, and mobile
Settings → AI providers had two implementations. oss/ee rendered the 0.112
connections table; `/m` hand-rolled the 0.111 design — a catalog of every
provider with "Configure now" per row, plus a separate OpenAI-compatible
endpoints table. Mobile users got no Add-provider catalog, no per-connection
naming, no model activation, and no Bedrock/Azure/Vertex.

Nothing forced the fork: the building blocks under @agenta/entity-ui/secretProvider
and @agenta/ui were already antd-free and portable. So move the page itself into
@agenta/settings-ui and let each host supply only its removal chrome — oss an
EnhancedModal, mobile a bottom sheet — which is the seam the package's other
tables already use.

The shared page uses DataTable rather than InfiniteVirtualTableFeatureShell, so
`/m` does not pull useStaticTable's table-scope atoms and virtualization. Cost on
desktop: column show/hide is per-mount instead of persisted, and the table no
longer paginates.

Also pass `docs` to SettingsPageShell on mobile. The helper and the entries both
existed and oss already called it, so every mobile settings tab was silently
dropping its docs link, not just this one.

Deletes SecretProviderTable and the two mobile sheets it drove. NamedSecretTable
stays — the Secrets tab still uses it.

Verified in the browser on /m and on EE: table, catalog, connection card with
model activation and harnesses, and delete confirmation with correct row identity.
